+++ main/playlist_tools/playlist_tools_menu_dyn_queries.js
@@ -1,7 +1,7 @@
 ﻿'use strict';
-//29/10/25
+//21/11/25
 
-/* global menusEnabled:readable, readmes:readable, menu:readable, menu_properties:readable, scriptName:readable, overwriteMenuProperties:readable, forcedQueryMenusEnabled:writable, defaultArgs:readable, disabledCount:writable, menuAltAllowed:readable, menuDisabled:readable, selectedFlags:readable, createSubMenuEditEntries:readable, libSearchMenu:readable */
+/* global menusEnabled:readable, readmes:readable, menu:readable, menu_properties:readable, scriptName:readable, overwriteMenuProperties:readable, forcedQueryMenusEnabled:writable, defaultArgs:readable, disabledCount:writable, menuAltAllowed:readable, menuDisabled:readable, selectedFlags:readable, createSubMenuEditEntries:readable, libSearchMenu:readable, entryMaxLength:readable */
 
 /* global MF_GRAYED:readable, MF_STRING:readable, folders:readable, _isFile:readable, globQuery:readable, globTags:readable, _qCond:readable, checkQuery:readable, isJSON:readable, WshShell:readable, popup:readable, queryJoin:readable, _ps:readable */
 
@@ -157,7 +157,7 @@
 										menu.newSeparator(entryMenuName);
 									} else {
 										// Create names for all entries
-										const queryName = (queryObj.name || '').cut(30);
+										const queryName = (queryObj.name || '').cut(entryMaxLength);
 										if (entryNames.has(queryName)) {
 											fb.ShowPopupMessage('There is an entry with duplicated name:\t' + queryName + '\nEdit the custom entries and either remove or rename it.\n\nEntry:\n' + JSON.stringify(queryObj, null, '\t'), scriptName + ': ' + name);
 											return;